refactor: 당첨 결과 계산, 처리 등을 수행한다는 의미에서 WinningProcessor로 클래스명 변경
